fix: fix website
fix the structure of the website
