+++ src/app/(customer)/layout.tsx
@@ -2,17 +2,14 @@ import { AuthContextProvider } from '@/components/features/auth/context'
 import { Footer } from '@/components/layouts/footer'
 import { Header } from '@/components/layouts/header'
 import { Navbar } from '@/components/layouts/navbar'
-import { Hero } from '@/components/layouts/hero'
 
-export default function HomePage() {
+export default function CustomerLayout({ children }: LayoutProps) {
   return (
     <AuthContextProvider>
       <Header>
         <Navbar />
       </Header>
-      <main className="flex grow">
-        <Hero />
-      </main>
+      <main className="flex grow">{children}</main>
       <Footer />
     </AuthContextProvider>
   )
