+++ src/app/(website)/whatsapp.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { FaWhatsapp } from 'react-icons/fa'
 
-export default function WhatsAppButton() {
+export const WhatsAppButton = () => {
   const phoneNumber = '4915257539557' // <-- change to your number (without +)
   const message = 'Hello, I am interested in your cars!'
