--- src/components/layouts/footer/conditional-footer.tsx
@@ -1,13 +1,0 @@
-'use client'
-
-import { usePathname } from 'next/navigation'
-import { Footer } from '@/components/layouts/footer'
-
-export function ConditionalFooter() {
-  const pathname = usePathname()
-  const excludeFooter = pathname.startsWith('/website')
-
-  if (excludeFooter) return null
-
-  return <Footer />
-}
